Move APDU commands specific to ACR122 to new class Fix typos in README
